Avoid base exception type error in cosmx reader
Changed line warning about discrepancy between fovs in labels and images from "raise logger.warning..." to just logger.warning because it causes a type error to "raise" something that doesn't inherit from the base exception class. This is the only instance of using "raise logger.warning" I was able to find in the repo, so I'm assuming this might be a typo.